Extract watchCacheHistory and watchCacheStorage to separate files
