+++ staging/src/k8s.io/apiserver/pkg/storage/cacher/delegator.go
@@ -85,7 +85,7 @@ func (c *CacheDelegator) EnableResourceSizeEstimation(keys storage.KeysFunc) err
 func (c *CacheDelegator) Delete(ctx context.Context, key string, out runtime.Object, preconditions *storage.Preconditions, validateDeletion storage.ValidateObjectFunc, cachedExistingObject runtime.Object, opts storage.DeleteOptions) error {
 	// Ignore the suggestion and try to pass down the current version of the object
 	// read from cache.
-	if elem, exists, err := c.cacher.watchCache.GetByKey(key); err != nil {
+	if elem, exists, err := c.cacher.watchCache.storage.GetByKey(key); err != nil {
 		klog.Errorf("GetByKey returned error: %v", err)
 	} else if exists {
 		// DeepCopy the object since we modify resource version when serializing the
@@ -205,7 +205,7 @@ func shouldDelegateListOnNotReadyCache(opts storage.ListOptions) bool {
 func (c *CacheDelegator) GuaranteedUpdate(ctx context.Context, key string, destination runtime.Object, ignoreNotFound bool, preconditions *storage.Preconditions, tryUpdate storage.UpdateFunc, cachedExistingObject runtime.Object) error {
 	// Ignore the suggestion and try to pass down the current version of the object
 	// read from cache.
-	if elem, exists, err := c.cacher.watchCache.GetByKey(key); err != nil {
+	if elem, exists, err := c.cacher.watchCache.storage.GetByKey(key); err != nil {
 		klog.Errorf("GetByKey returned error: %v", err)
 	} else if exists {
 		// DeepCopy the object since we modify resource version when serializing the
