+++ staging/src/k8s.io/apiserver/pkg/storage/cacher/watch_cache_test.go
@@ -212,7 +212,7 @@ func TestWatchCacheBasic(t *testing.T) {
 	if err := s.Add(pod1); err != nil {
 		t.Errorf("unexpected error: %v", err)
 	}
-	if item, ok, _ := s.Get(pod1); !ok {
+	if item, ok, _ := s.storage.Get(pod1); !ok {
 		t.Errorf("didn't find pod")
 	} else {
 		expected := makeTestStoreElement(makeTestPod("pod", 1))
@@ -224,7 +224,7 @@ func TestWatchCacheBasic(t *testing.T) {
 	if err := s.Update(pod2); err != nil {
 		t.Errorf("unexpected error: %v", err)
 	}
-	if item, ok, _ := s.Get(pod2); !ok {
+	if item, ok, _ := s.storage.Get(pod2); !ok {
 		t.Errorf("didn't find pod")
 	} else {
 		expected := makeTestStoreElement(makeTestPod("pod", 2))
@@ -236,7 +236,7 @@ func TestWatchCacheBasic(t *testing.T) {
 	if err := s.Delete(pod3); err != nil {
 		t.Errorf("unexpected error: %v", err)
 	}
-	if _, ok, _ := s.Get(pod3); ok {
+	if _, ok, _ := s.storage.Get(pod3); ok {
 		t.Errorf("found pod")
 	}
 
@@ -251,7 +251,7 @@ func TestWatchCacheBasic(t *testing.T) {
 			"/prefix/ns/pod3": *makeTestStoreElement(makeTestPod("pod3", 6)),
 		}
 		items := make(map[string]store.Element)
-		for _, item := range s.List() {
+		for _, item := range s.storage.List() {
 			elem := item.(*store.Element)
 			items[elem.Key] = *elem
 		}
@@ -271,7 +271,7 @@ func TestWatchCacheBasic(t *testing.T) {
 			"/prefix/ns/pod5": *makeTestStoreElement(makeTestPod("pod5", 8)),
 		}
 		items := make(map[string]store.Element)
-		for _, item := range s.List() {
+		for _, item := range s.storage.List() {
 			elem := item.(*store.Element)
 			items[elem.Key] = *elem
 		}
